Create separate ViewModel files, organize folder structure
