+++ TuneIn/Firebase/AuthenticationManager.swift
