+++ TuneIn/Views/SettingsView.swift
@@ -8,30 +8,6 @@
 import SwiftUI
 import PhotosUI
 
-enum ImageSaveError: Error {
-    case dataLoadingFailed
-}
-
-
-@MainActor
-final class SettingsViewModel: ObservableObject {
-    func signOut() throws {
-        try AuthenticationManager.shared.signOut()
-    }
-    
-    func updatePassword(password: String) async throws {
-        try await AuthenticationManager.shared.updatePassword(password: password)
-    }
-    
-    func saveProfileImage(item: PhotosPickerItem, userId: String) async throws -> URL {
-        guard let data = try await item.loadTransferable(type: Data.self) else {
-            throw ImageSaveError.dataLoadingFailed
-        }
-        let (path, _) = try await StorageManager.shared.saveImage(data: data, userId: userId)
-        return try await StorageManager.shared.getImageUrl(userId: userId, path: path)
-    }
-}
-
 struct SettingsView: View {
     @EnvironmentObject var userModel: UserModel
     @Environment(\.dismiss) var dismiss
